Consolidate ITerminalCommand interface
Part of #181538
